Se hicieron cambios en las BLL, Contexto y algunos registros para arreglar varios errores, se hizo una migracion, se programacion las clases de los registros y se agrego un mensaje de error a la aplicaccion.

+++ Entidades/PrestamosDetalle.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Text;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
-using System.Text;
 
 namespace PrestamosdeJuegos.Entidades
 {
@@ -13,7 +13,7 @@ namespace PrestamosdeJuegos.Entidades
         public int PrestamoDetalleId { get; set; }
         public int PrestamosId { get; set; }
         public int JuegoId { get; set; }
-        public int CantidadJuegos { get; set; }
+        public double CantidadJuegos { get; set; }
 
         [ForeignKey("JuegoId")]
         public virtual Juegos FK_Juegos { get; set; }
